models/owner: Convert `OwnerKind` to `pg_enum` (#7308)

+++ src/admin/transfer_crates.rs
@@ -60,7 +60,7 @@ fn transfer(opts: Opts, conn: &mut PgConnection) {
 
     let crate_owners = crate_owners::table
         .filter(crate_owners::owner_id.eq(from.id))
-        .filter(crate_owners::owner_kind.eq(OwnerKind::User as i32));
+        .filter(crate_owners::owner_kind.eq(OwnerKind::User));
     let crates: Vec<Crate> = Crate::all()
         .filter(crates::id.eq_any(crate_owners.select(crate_owners::crate_id)))
         .load(conn)
